requirements.txt
Pillow>=10.0.0
ImageHash>=4.3.1
opencv-python>=4.8.0
